crawler/requirements.txt
feedparser==6.0.11
newspaper3k==0.2.8
httpx==0.27.0
openai==1.35.0
beautifulsoup4==4.12.3
schedule==1.2.2
lxml==5.2.2
Pillow==10.3.0
